Move fetchUsersAndUsernames to userSlice
Which allows us to get rid of userThunks.ts
